feat(node/e2e-test): add peer count test
