--- tests/node/opp2p_rpc.go
@@ -1,1 +1,0 @@
-package node
